+++ src/modules/index.js
@@ -2,7 +2,7 @@ import { combineReducers } from "redux";
 import user from "./user";
 
 const rootReducer = combineReducers({
-  user,
+  user
 })
 
 export default rootReducer;
