+++ src/pages/layout/Layout.jsx
@@ -47,7 +47,6 @@ const Layout = () => {
           }));
           dispatch(setUserStatus(false));
           localStorage.clear();
-          return;
         }
         
         const datas = await response.json();
@@ -58,9 +57,10 @@ const Layout = () => {
       };
       getUserDatas()
     }
-  }, [jwtToken]); 
+  }, [localJwtToken]); 
 
-  console.log("currentUser", currentUser)
+  console.log("리덕스 유저", currentUser)
+  console.log("리덕스 유저 상태", isLogin)
 
   const handleLogout = () => {
     localStorage.clear();
